fix(deps): bump all dependencies, fixing aws-lc-sys build failure (#90)
Update Cargo.lock and bump breaking dependency versions:
- rand 0.8 → 0.9 (thread_rng→rng, OsRng fill_bytes→try_fill_bytes)
- fastcdc 3 → 4 (constants changed from u32 to usize)
- russh 0.58 → 0.60, rusty-s3 0.8 → 0.9, lru 0.16 → 0.17, tray-icon 0.21 → 0.22
- aws-lc-sys 0.38 → 0.40 fixes CFLAGS -O2 jitterentropy conflict

Closes #90

+++ crates/vykar-core/src/chunker/mod.rs
@@ -7,11 +7,21 @@ use crate::config::ChunkerConfig;
 /// Chunk a byte slice using FastCDC content-defined chunking.
 /// Returns a vector of `(offset, length)` pairs.
 pub fn chunk_data(data: &[u8], config: &ChunkerConfig) -> Vec<(usize, usize)> {
-    let chunker = FastCDC::new(data, config.min_size, config.avg_size, config.max_size);
+    let chunker = FastCDC::new(
+        data,
+        config.min_size as usize,
+        config.avg_size as usize,
+        config.max_size as usize,
+    );
     chunker.map(|chunk| (chunk.offset, chunk.length)).collect()
 }
 
 /// Chunk a reader stream using FastCDC content-defined chunking.
 pub fn chunk_stream<R: Read>(source: R, config: &ChunkerConfig) -> StreamCDC<R> {
-    StreamCDC::new(source, config.min_size, config.avg_size, config.max_size)
+    StreamCDC::new(
+        source,
+        config.min_size as usize,
+        config.avg_size as usize,
+        config.max_size as usize,
+    )
 }
